requirements.txt
apache_beam[gcp]==2.55.0
psycopg2-binary==2.9.9
google-cloud-storage
pyarrow
setuptools
apache_beam[dataframe]
cx_Oracle
pandas
fsspec
gcsfs